shell(theme): reject unknown flags + pay layer-back-edge debt (#2255)
theme list/apply/export silently ignored stray dash-flags and exited 0,
which is false confidence. Parse via shared parseKnownFlags (any position,
honours --), fail unknown flags with exit 1, and move theme presets into
base/ so the shell command no longer imports ui/.
