+++ packages/dev-tools/tools/layer-back-edge-baseline.json
@@ -26,7 +26,6 @@
   "packages/webapp/src/shell/supplemental-commands/secret-backends.ts": 2,
   "packages/webapp/src/shell/supplemental-commands/serial-command.ts": 1,
   "packages/webapp/src/shell/supplemental-commands/serve-command.ts": 3,
-  "packages/webapp/src/shell/supplemental-commands/theme-command.ts": 3,
   "packages/webapp/src/shell/supplemental-commands/upskill/tabs.ts": 1,
   "packages/webapp/src/shell/supplemental-commands/usb-command.ts": 1
 }
