+++ packages/webapp/src/shell/supplemental-commands/theme-command.ts
@@ -1,9 +1,14 @@
 import type { Command } from 'just-bash';
 import { defineCommand } from 'just-bash';
 import { getPanelRpcClient } from '../../kernel/panel-rpc.js';
+import { parseKnownFlags } from './subcommand-flags.js';
+import { isHelpRequest, subcommandHelpText } from './subcommand-help.js';
 
 type CommandContext = Parameters<Parameters<typeof defineCommand>[1]>[1];
 type CommandResult = { stdout: string; stderr: string; exitCode: number };
+
+/** Theme owns no value-taking flags; keep the empty list next to isHelpRequest. */
+const THEME_VALUE_FLAGS: readonly string[] = [];
 
 const HELP = `usage: theme <subcommand> [args]
 
@@ -16,17 +21,30 @@ Subcommands:
   export <id> <path>   Export a theme to a VFS path
 `;
 
+function fail(sub: string, message: string): CommandResult {
+  return { stdout: '', stderr: `theme ${sub}: ${message}\n`, exitCode: 1 };
+}
+
 export function createThemeCommand(): Command {
   return defineCommand('theme', async (args, ctx) => {
-    if (args.length === 0 || args.includes('--help') || args.includes('-h')) {
+    if (args.length === 0 || args[0] === '--help' || args[0] === '-h') {
       return { stdout: HELP, stderr: '', exitCode: 0 };
     }
 
     const sub = args[0];
+    // Help before the handler so `theme apply --help` never treats `--help`
+    // as a theme id. `theme apply -- --help` still uses the literal path.
+    if (isHelpRequest(args.slice(1), { valueFlags: THEME_VALUE_FLAGS })) {
+      return {
+        stdout: subcommandHelpText('theme', sub, HELP),
+        stderr: '',
+        exitCode: 0,
+      };
+    }
 
-    if (sub === 'list') return listThemes();
-    if (sub === 'current') return currentTheme();
-    if (sub === 'reset') return resetTheme();
+    if (sub === 'list') return listThemes(args);
+    if (sub === 'current') return currentTheme(args);
+    if (sub === 'reset') return resetTheme(args);
     if (sub === 'apply') return applyTheme(args.slice(1), ctx);
     if (sub === 'export') return exportThemeCmd(args.slice(1), ctx);
 
@@ -34,8 +52,13 @@ export function createThemeCommand(): Command {
   });
 }
 
-async function listThemes(): Promise<CommandResult> {
-  const { PRESETS } = await import('../../ui/theme-presets.js');
+async function listThemes(args: string[]): Promise<CommandResult> {
+  const parsed = parseKnownFlags(args.slice(1), { value: THEME_VALUE_FLAGS });
+  if ('error' in parsed) return fail('list', parsed.error);
+
+  // Dynamic import keeps the preset table out of the kernel-worker first-load
+  // graph; importing from `base/` (not `ui/`) pays the layer-back-edge debt.
+  const { PRESETS } = await import('../../base/theme-presets.js');
   const lines: string[] = ['Presets:'];
   for (const p of PRESETS) {
     lines.push(`  ${p.id.padEnd(16)} ${p.name}`);
@@ -44,7 +67,10 @@ async function listThemes(): Promise<CommandResult> {
   return { stdout: lines.join('\n'), stderr: '', exitCode: 0 };
 }
 
-async function currentTheme(): Promise<CommandResult> {
+async function currentTheme(args: string[]): Promise<CommandResult> {
+  const parsed = parseKnownFlags(args.slice(1), { value: THEME_VALUE_FLAGS });
+  if ('error' in parsed) return fail('current', parsed.error);
+
   return {
     stdout: 'Check the Theme settings dialog for the active theme.\n',
     stderr: '',
@@ -52,7 +78,10 @@ async function currentTheme(): Promise<CommandResult> {
   };
 }
 
-async function resetTheme(): Promise<CommandResult> {
+async function resetTheme(args: string[]): Promise<CommandResult> {
+  const parsed = parseKnownFlags(args.slice(1), { value: THEME_VALUE_FLAGS });
+  if ('error' in parsed) return fail('reset', parsed.error);
+
   const panelRpc = getPanelRpcClient();
   if (!panelRpc) {
     return { stdout: '', stderr: 'theme: no panel-RPC connection available\n', exitCode: 1 };
@@ -62,7 +91,10 @@ async function resetTheme(): Promise<CommandResult> {
 }
 
 async function applyTheme(args: string[], ctx: CommandContext): Promise<CommandResult> {
-  if (args.length === 0) {
+  const parsed = parseKnownFlags(args, { value: THEME_VALUE_FLAGS });
+  if ('error' in parsed) return fail('apply', parsed.error);
+
+  if (parsed.positionals.length === 0) {
     return { stdout: '', stderr: 'theme apply: missing theme id or path\n', exitCode: 1 };
   }
 
@@ -71,10 +103,9 @@ async function applyTheme(args: string[], ctx: CommandContext): Promise<CommandR
     return { stdout: '', stderr: 'theme: no panel-RPC connection available\n', exitCode: 1 };
   }
 
-  const target = args[0];
+  const target = parsed.positionals[0];
 
-  // Check if it's a known preset id
-  const { PRESETS } = await import('../../ui/theme-presets.js');
+  const { PRESETS } = await import('../../base/theme-presets.js');
   const preset = PRESETS.find((p) => p.id === target);
   if (preset) {
     const json = JSON.stringify(preset);
@@ -97,8 +128,8 @@ async function applyTheme(args: string[], ctx: CommandContext): Promise<CommandR
 
   // Validate JSON structure before sending
   try {
-    const parsed = JSON.parse(content);
-    if (!parsed.id || !parsed.name || !parsed.base || !parsed.tokens) {
+    const parsedTheme = JSON.parse(content);
+    if (!parsedTheme.id || !parsedTheme.name || !parsedTheme.base || !parsedTheme.tokens) {
       return {
         stdout: '',
         stderr: 'theme apply: invalid theme file (missing id, name, base, or tokens)\n',
@@ -114,12 +145,15 @@ async function applyTheme(args: string[], ctx: CommandContext): Promise<CommandR
 }
 
 async function exportThemeCmd(args: string[], ctx: CommandContext): Promise<CommandResult> {
-  if (args.length < 2) {
+  const parsed = parseKnownFlags(args, { value: THEME_VALUE_FLAGS });
+  if ('error' in parsed) return fail('export', parsed.error);
+
+  if (parsed.positionals.length < 2) {
     return { stdout: '', stderr: 'theme export: usage: theme export <id> <path>\n', exitCode: 1 };
   }
 
-  const [id, path] = args;
-  const { PRESETS } = await import('../../ui/theme-presets.js');
+  const [id, path] = parsed.positionals;
+  const { PRESETS } = await import('../../base/theme-presets.js');
   const preset = PRESETS.find((p) => p.id === id);
   if (!preset) {
     return { stdout: '', stderr: `theme export: unknown preset id "${id}"\n`, exitCode: 1 };
