+++ packages/webapp/src/ui/theme-presets.ts
@@ -1,263 +1,6 @@
-// packages/webapp/src/ui/theme-presets.ts
-import type { SliccTheme } from './theme-types.js';
-
-export const PRESETS: SliccTheme[] = [
-  {
-    id: 'vanilla',
-    name: 'Vanilla',
-    base: 'light',
-    disableShader: true,
-    tokens: {
-      '--s2-gray-25': '#fffdf8',
-      '--s2-gray-50': '#faf6f0',
-      '--s2-gray-75': '#f5f0e8',
-      '--s2-gray-100': '#ede7dc',
-      '--s2-gray-200': '#e0d8cc',
-      '--s2-gray-300': '#d4cbbf',
-      '--s2-gray-400': '#b8ad9e',
-      '--s2-gray-500': '#8c7f6f',
-      '--s2-gray-600': '#6b5f50',
-      '--s2-gray-700': '#4d4235',
-      '--s2-gray-800': '#2e2418',
-      '--s2-gray-900': '#1a1008',
-      '--s2-gray-1000': '#000000',
-      '--s2-bg-base': '#fffdf8',
-      '--s2-bg-layer-1': '#faf6f0',
-      '--s2-bg-layer-2': '#f5f0e8',
-      '--s2-bg-elevated': '#ede7dc',
-      '--s2-bg-sunken': '#f5f0e8',
-      '--s2-accent': '#a0522d',
-      '--s2-accent-hover': '#8b4513',
-      '--s2-accent-down': '#6b3410',
-      '--slicc-cone': '#d2691e',
-      '--s2-border-default': '#d4cbbf',
-      '--s2-border-subtle': '#e0d8cc',
-      '--canvas': '#fffdf8',
-      '--bg': '#f5f0e8',
-      '--ghost': '#f5f0e8',
-      '--desk': '#f5f0e8',
-      '--ink': '#1a1008',
-      '--deep': '#1a1008',
-      '--txt-2': '#6b5f50',
-      '--txt-3': '#8c7f6f',
-      '--line': '#d4cbbf',
-      '--ctx': '#a0522d',
-      '--shaderbg': '#fffdf8',
-    },
-  },
-  {
-    id: 'midnight-scoop',
-    name: 'Midnight Scoop',
-    base: 'dark',
-    disableShader: true,
-    tokens: {
-      '--s2-gray-25': '#0d1117',
-      '--s2-gray-50': '#131a24',
-      '--s2-gray-75': '#1a2332',
-      '--s2-gray-100': '#212d40',
-      '--s2-gray-200': '#2d3b50',
-      '--s2-gray-300': '#3a4a63',
-      '--s2-gray-400': '#4d5f7a',
-      '--s2-gray-500': '#6b7f99',
-      '--s2-gray-600': '#8899b3',
-      '--s2-gray-700': '#a8b8d0',
-      '--s2-gray-800': '#cdd8e8',
-      '--s2-gray-900': '#e8eef5',
-      '--s2-gray-1000': '#ffffff',
-      '--s2-bg-base': '#0d1117',
-      '--s2-bg-layer-1': '#131a24',
-      '--s2-bg-layer-2': '#1a2332',
-      '--s2-bg-elevated': '#212d40',
-      '--s2-bg-sunken': '#080c12',
-      '--s2-accent': '#58a6ff',
-      '--s2-accent-hover': '#79b8ff',
-      '--s2-accent-down': '#3d8bda',
-      '--slicc-cone': '#f0883e',
-      '--s2-border-default': '#3a4a63',
-      '--s2-border-subtle': '#2d3b50',
-      '--canvas': '#0d1117',
-      '--bg': '#080c12',
-      '--ghost': '#1a2332',
-      '--desk': '#1a2332',
-      '--ink': '#e8eef5',
-      '--deep': '#e8eef5',
-      '--txt-2': '#8899b3',
-      '--txt-3': '#6b7f99',
-      '--line': '#3a4a63',
-      '--ctx': '#58a6ff',
-      '--shaderbg': '#0d1117',
-    },
-  },
-  {
-    id: 'matcha-float',
-    name: 'Matcha Float',
-    base: 'dark',
-    disableShader: true,
-    tokens: {
-      '--s2-gray-25': '#0f1a14',
-      '--s2-gray-50': '#14221b',
-      '--s2-gray-75': '#1a2d23',
-      '--s2-gray-100': '#21382c',
-      '--s2-gray-200': '#2d4a3a',
-      '--s2-gray-300': '#3a5c48',
-      '--s2-gray-400': '#4d7360',
-      '--s2-gray-500': '#6b9480',
-      '--s2-gray-600': '#88b09c',
-      '--s2-gray-700': '#a8c8b8',
-      '--s2-gray-800': '#cce0d5',
-      '--s2-gray-900': '#e8f2ec',
-      '--s2-gray-1000': '#ffffff',
-      '--s2-bg-base': '#0f1a14',
-      '--s2-bg-layer-1': '#14221b',
-      '--s2-bg-layer-2': '#1a2d23',
-      '--s2-bg-elevated': '#21382c',
-      '--s2-bg-sunken': '#0a120e',
-      '--s2-accent': '#6bce9a',
-      '--s2-accent-hover': '#88dbb0',
-      '--s2-accent-down': '#4fb882',
-      '--slicc-cone': '#a8d86c',
-      '--s2-border-default': '#3a5c48',
-      '--s2-border-subtle': '#2d4a3a',
-      '--canvas': '#0f1a14',
-      '--bg': '#0a120e',
-      '--ghost': '#1a2d23',
-      '--desk': '#1a2d23',
-      '--ink': '#e8f2ec',
-      '--deep': '#e8f2ec',
-      '--txt-2': '#88b09c',
-      '--txt-3': '#6b9480',
-      '--line': '#3a5c48',
-      '--ctx': '#6bce9a',
-      '--shaderbg': '#0f1a14',
-    },
-  },
-  {
-    id: 'berry-cone',
-    name: 'Berry Cone',
-    base: 'dark',
-    disableShader: true,
-    tokens: {
-      '--s2-gray-25': '#150a1a',
-      '--s2-gray-50': '#1e1024',
-      '--s2-gray-75': '#281830',
-      '--s2-gray-100': '#33203d',
-      '--s2-gray-200': '#44304f',
-      '--s2-gray-300': '#574063',
-      '--s2-gray-400': '#6e5580',
-      '--s2-gray-500': '#8c7099',
-      '--s2-gray-600': '#a88cb3',
-      '--s2-gray-700': '#c0a8cc',
-      '--s2-gray-800': '#d8c8e0',
-      '--s2-gray-900': '#f0e8f4',
-      '--s2-gray-1000': '#ffffff',
-      '--s2-bg-base': '#150a1a',
-      '--s2-bg-layer-1': '#1e1024',
-      '--s2-bg-layer-2': '#281830',
-      '--s2-bg-elevated': '#33203d',
-      '--s2-bg-sunken': '#0f0614',
-      '--s2-accent': '#e06be0',
-      '--s2-accent-hover': '#e88ae8',
-      '--s2-accent-down': '#c84ec8',
-      '--slicc-cone': '#ff6b9d',
-      '--s2-border-default': '#574063',
-      '--s2-border-subtle': '#44304f',
-      '--canvas': '#150a1a',
-      '--bg': '#0f0614',
-      '--ghost': '#281830',
-      '--desk': '#281830',
-      '--ink': '#f0e8f4',
-      '--deep': '#f0e8f4',
-      '--txt-2': '#a88cb3',
-      '--txt-3': '#8c7099',
-      '--line': '#574063',
-      '--ctx': '#e06be0',
-      '--shaderbg': '#150a1a',
-    },
-  },
-  {
-    id: 'caramel-swirl',
-    name: 'Caramel Swirl',
-    base: 'light',
-    disableShader: true,
-    tokens: {
-      '--s2-gray-25': '#fdf9f3',
-      '--s2-gray-50': '#f8f2e8',
-      '--s2-gray-75': '#f2eadd',
-      '--s2-gray-100': '#e8dfd0',
-      '--s2-gray-200': '#ddd2c0',
-      '--s2-gray-300': '#d0c3b0',
-      '--s2-gray-400': '#b8a890',
-      '--s2-gray-500': '#8f7a60',
-      '--s2-gray-600': '#6b5a42',
-      '--s2-gray-700': '#4d3f2a',
-      '--s2-gray-800': '#2e2515',
-      '--s2-gray-900': '#1a1408',
-      '--s2-gray-1000': '#000000',
-      '--s2-bg-base': '#fdf9f3',
-      '--s2-bg-layer-1': '#f8f2e8',
-      '--s2-bg-layer-2': '#f2eadd',
-      '--s2-bg-elevated': '#e8dfd0',
-      '--s2-bg-sunken': '#f2eadd',
-      '--s2-accent': '#c87830',
-      '--s2-accent-hover': '#b06828',
-      '--s2-accent-down': '#985820',
-      '--slicc-cone': '#e89040',
-      '--s2-border-default': '#d0c3b0',
-      '--s2-border-subtle': '#ddd2c0',
-      '--canvas': '#fdf9f3',
-      '--bg': '#f2eadd',
-      '--ghost': '#f2eadd',
-      '--desk': '#f2eadd',
-      '--ink': '#1a1408',
-      '--deep': '#1a1408',
-      '--txt-2': '#6b5a42',
-      '--txt-3': '#8f7a60',
-      '--line': '#d0c3b0',
-      '--ctx': '#c87830',
-      '--shaderbg': '#fdf9f3',
-    },
-  },
-  {
-    id: 'sorbet',
-    name: 'Sorbet',
-    base: 'light',
-    disableShader: true,
-    tokens: {
-      '--s2-gray-25': '#fff8f6',
-      '--s2-gray-50': '#fef0ed',
-      '--s2-gray-75': '#fce8e3',
-      '--s2-gray-100': '#f8ddd6',
-      '--s2-gray-200': '#f0d0c8',
-      '--s2-gray-300': '#e8c0b8',
-      '--s2-gray-400': '#d4a89e',
-      '--s2-gray-500': '#a8807a',
-      '--s2-gray-600': '#80605c',
-      '--s2-gray-700': '#5c4240',
-      '--s2-gray-800': '#3a2826',
-      '--s2-gray-900': '#1e1412',
-      '--s2-gray-1000': '#000000',
-      '--s2-bg-base': '#fff8f6',
-      '--s2-bg-layer-1': '#fef0ed',
-      '--s2-bg-layer-2': '#fce8e3',
-      '--s2-bg-elevated': '#f8ddd6',
-      '--s2-bg-sunken': '#fce8e3',
-      '--s2-accent': '#e86050',
-      '--s2-accent-hover': '#d04840',
-      '--s2-accent-down': '#b83830',
-      '--slicc-cone': '#ff7860',
-      '--s2-border-default': '#e8c0b8',
-      '--s2-border-subtle': '#f0d0c8',
-      '--canvas': '#fff8f6',
-      '--bg': '#fce8e3',
-      '--ghost': '#fce8e3',
-      '--desk': '#fce8e3',
-      '--ink': '#1e1412',
-      '--deep': '#1e1412',
-      '--txt-2': '#80605c',
-      '--txt-3': '#a8807a',
-      '--line': '#e8c0b8',
-      '--ctx': '#e86050',
-      '--shaderbg': '#fff8f6',
-    },
-  },
-];
+/**
+ * Re-export preset themes from `base/` so existing `ui/` imports keep working.
+ * The data lives in the base layer so `theme` (shell) can read it without a
+ * shell → ui layer back-edge (#2255 boy-scout).
+ */
+export { PRESETS } from '../base/theme-presets.js';
